+++ src/convert/fill.rs
@@ -3,7 +3,10 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 // external
-use svgdom;
+use svgdom::{
+    self,
+    PaintFallback,
+};
 
 // self
 use tree;
@@ -15,42 +18,7 @@ pub fn convert(
     rtree: &tree::Tree,
     attrs: &svgdom::Attributes,
 ) -> Option<tree::Fill> {
-    let paint = if let Some(fill) = attrs.get_type(AId::Fill) {
-        match *fill {
-            AValue::Color(c) => {
-                tree::Paint::Color(c)
-            }
-            AValue::FuncLink(ref link) => {
-                // a-fill-016.svg
-                // a-fill-017.svg
-                // a-fill-018.svg
-
-                let mut p = None;
-                if link.is_gradient() || link.is_tag_name(EId::Pattern) {
-                    if let Some(node) = rtree.defs_by_id(&link.id()) {
-                        p = Some(tree::Paint::Link(node.id().to_string()));
-                    }
-                } else {
-                    // a-fill-023.svg
-                    warn!("'{}' cannot be used for filling.",
-                          link.tag_name());
-                    return None;
-                }
-
-                p?
-            }
-            AValue::None => {
-                // a-fill-020.svg
-                return None;
-            }
-            _ => {
-                warn!("An invalid fill value: {}. Skipped.", fill);
-                return None;
-            }
-        }
-    } else {
-        return None;
-    };
+    let paint = resolve_paint(rtree, attrs, AId::Fill)?;
 
     let fill_opacity = attrs.get_number(AId::FillOpacity).unwrap_or(1.0);
 
@@ -67,4 +35,44 @@ pub fn convert(
     };
 
     Some(fill)
+}
+
+pub fn resolve_paint(
+    rtree: &tree::Tree,
+    attrs: &svgdom::Attributes,
+    aid: AId,
+) -> Option<tree::Paint> {
+    match attrs.get_type(aid) {
+        Some(AValue::Color(c)) => {
+            Some(tree::Paint::Color(*c))
+        }
+        Some(AValue::Paint(ref link, fallback)) => {
+            // a-fill-016.svg
+            // a-fill-017.svg
+            // a-fill-018.svg
+
+            if link.is_gradient() || link.is_tag_name(EId::Pattern) {
+                if let Some(node) = rtree.defs_by_id(&link.id()) {
+                    Some(tree::Paint::Link(node.id().to_string()))
+                } else if let Some(PaintFallback::Color(c)) = fallback {
+                    Some(tree::Paint::Color(*c))
+                } else {
+                    None
+                }
+            } else {
+                // a-fill-023.svg
+                warn!("'{}' cannot be used to {} the shape.", link.tag_name(), aid);
+                None
+            }
+        }
+        Some(AValue::None) => {
+            // a-fill-020.svg
+            None
+        }
+        Some(av) => {
+            warn!("An invalid {} value: {}. Skipped.", aid, av);
+            None
+        }
+        None => None,
+    }
 }